add and update apis

+++ sweep/src/services/calendar/calendars.ts
@@ -18,6 +18,20 @@ export async function getCalendar(id: string) {
   }
 }
 
+export async function getCalendarData(id: number, month: string) {
+  try {
+    const response = await axios.get(`/v1/calendars/${id}/schedules/`, {
+      params: {
+        month,
+      },
+    });
+
+    return response.data;
+  } catch {
+    return null;
+  }
+}
+
 export async function createCalendar() {
   try {
     const response = await axios.post("/v1/calendars/");
@@ -27,7 +41,11 @@ export async function createCalendar() {
   }
 }
 
-export async function deleteCalendar(id: string) {
+export async function deleteCalendar(id: number | undefined) {
+  if (!id) {
+    return null;
+  }
+  
   try {
     await axios.delete(`/v1/calendars/${id}/`);
     return true;
@@ -49,7 +67,11 @@ export async function updateCalendarInfo(id: number, name: string, color: string
   }
 }
 
-export async function toggleCalendarNotification(id: string) {
+export async function toggleCalendarNotification(id: number | undefined) {
+  if (!id) {
+    return null;
+  }
+
   try {
     const response = await axios.patch(`/v1/calendars/${id}/notification/`);
     return response.data;
@@ -58,7 +80,11 @@ export async function toggleCalendarNotification(id: string) {
   }
 }
 
-export async function toggleCalendarDaily(id: string, time: string) {
+export async function toggleCalendarDaily(id: number | undefined, time: string) {
+  if (!id) {
+    return null;
+  }
+
   try {
     const response = await axios.patch(`/v1/calendars/${id}/daily/`, {
       time,
